chore(security): Pin pr-title-checker.yml to commit
This PR pins the `pr-title-checker` third-party action to a full-length commit SHA — that of release 1.4.0.

Pinning an action to a full-length commit SHA is currently the only way to use an action as an immutable release. Pinning to a particular SHA helps mitigate the risk of a bad actor adding a backdoor to the action's repository.